BE-FM-S11P12A606-338-1: disallow deleting default image file

+++ src/main/java/com/example/hotsix/controller/recruit/RecruitController.java
@@ -130,7 +130,7 @@ public class RecruitController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/teambuilding/recruit/{id}")
     public void deleteRecruit(@PathVariable("id") Recruit recruit) throws IOException {
-        storageService.remove(recruit.getThumbnail());
+        if (!recruit.isDefaultThumbnail()) storageService.remove(recruit.getThumbnail());
 
         recruitService.delete(recruit);
     }
